Move imputation function to separate module
